requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic_settings
passlib[bcrypt]
bcrypt==4.0.1
python-jose
authlib
pydantic[email]
python-multipart
httpx